+++ backend/controllers/userController.js
@@ -1,16 +1,19 @@
 const User = require("../models/user");
+const bcrypt = require("bcrypt");
+const jwt = require("jsonwebtoken");
 
-// Get all users
-const getAllUsers = async (req, res) => {
-  try {
-    const users = await User.find();
-    res.json(users);
-  } catch (error) {
-    res.status(500).json({ error: "Internal Server Error" });
-  }
+const saltRounds = 10;
+const accessTokenSecret = "storymultiverse-access-token-secret";
+
+// Helper function to generate an access token
+const generateAccessToken = (userId, username, tier, membershipStatus) => {
+  return jwt.sign(
+    { userId, username, tier, membershipStatus },
+    accessTokenSecret
+  );
 };
 
-// Get a user by ID
+// Get user by ID
 const getUserById = async (req, res) => {
   try {
     const userId = req.params.id;
@@ -26,20 +29,65 @@ const getUserById = async (req, res) => {
   }
 };
 
-// Create a new user
+// Create new user (Register)
 const createUser = async (req, res) => {
   try {
-    const newUser = req.body;
-    const createdUser = await User.create(newUser);
+    const { userEmail, password } = req.body;
 
-    res.status(201).json(createdUser);
+    if (password.length < 8) {
+      return res
+        .status(400)
+        .json({ error: "Password should be at least 8 characters long" });
+    }
+
+    const existingUser = await User.findOne({ userEmail });
+    if (existingUser) {
+      return res.status(400).json({ error: "Email already registered" });
+    }
+
+    const hashedPassword = await bcrypt.hash(password, saltRounds);
+
+    const newUser = await User.create({
+      ...req.body,
+      password: hashedPassword,
+    });
+
+    res.status(201).json({ message: "Registration successful", user: newUser });
   } catch (error) {
     res.status(500).json({ error: "Internal Server Error" });
   }
 };
 
-// Update a user
+// User Login
+const loginUser = async (req, res) => {
+  try {
+    const { userEmail, password } = req.body;
+
+    const user = await User.findOne({ userEmail });
+    if (!user) {
+      return res.status(401).json({ error: "Invalid credentials" });
+    }
+
+    const isMatch = await bcrypt.compare(password, user.password);
+    if (!isMatch) {
+      return res.status(401).json({ error: "Invalid credentials" });
+    }
+
+    const accessToken = generateAccessToken(
+      user._id,
+      user.name,
+      user.tier,
+      user.membershipStatus
+    );
+    res.json({ accessToken });
+  } catch (error) {
+    res.status(500).json({ error: "Internal Server Error" });
+  }
+};
+
+// Update user
 const updateUser = async (req, res) => {
+  console.log(res);
   try {
     const userId = req.params.id;
     const updatedUser = req.body;
@@ -56,7 +104,7 @@ const updateUser = async (req, res) => {
   }
 };
 
-// Delete a user
+// Delete user
 const deleteUser = async (req, res) => {
   try {
     const userId = req.params.id;
@@ -72,10 +120,22 @@ const deleteUser = async (req, res) => {
   }
 };
 
+const requestPasswordReset = async (req, res) => {
+  // TODO: Implement the request password reset logic here
+  res.status(501).json({ message: "Not implemented yet" });
+};
+
+const resetPassword = async (req, res) => {
+  // TODO: Implement the reset password logic here using the provided token
+  res.status(501).json({ message: "Not implemented yet" });
+};
+
 module.exports = {
-  getAllUsers,
   getUserById,
   createUser,
+  loginUser, // Added loginUser function
   updateUser,
   deleteUser,
+  requestPasswordReset,
+  resetPassword,
 };