+++ frontend/src/pages/Login.js
@@ -14,7 +14,7 @@ export default function Login() {
       <Navbar />
       <div className="container">
         <h1>Unlock the Multiverse</h1>
-        <h2>sign-in or sign-up embark on the Storymultiverse adventure!</h2>
+        <h2>sign-in or sign-up & embark on the Storymultiverse adventure!</h2>
         <UserLogin />
       </div>
       <FooterComponent />