+++ frontend/src/pages/Register.js
@@ -3,6 +3,7 @@ import React from "react";
 //******* Import component(s) *******/
 import FooterComponent from "../components/FooterComponent";
 import Navbar from "../components/NavBar";
+import UserRegister from "../components/UserRegister";
 
 //******* Import style logic *******/
 //Here
@@ -13,7 +14,8 @@ export default function Register() {
       <Navbar />
       <div className="container">
         <h1>Register page</h1>
-        <h2>Here is your support plateform, we are here to help you!</h2>
+        <h2>Create an account in a single step !</h2>
+        <UserRegister />
       </div>
       <FooterComponent />
     </div>