+++ CHANGELOG.md
@@ -1,23 +1,77 @@
 # Changelog
 
-## UI [1.10.1] 2023-08-07
+## UI [1.11.2] 2023-08-11
 
 ### Changed
-- CSS Styling updated for contact us page
-- Contact details added
-- Map added 
+
+- clean up footer, contact register , login
+- link login icon to login page
+- removed uderline deco on title (css) as it was applicable to entire app
+
+## UI [1.11.1] 2023-08-11
+
+### Changed
+
+- pages: removed surprised me
+- removed services folder as it is not required anymore
+- Removed unsused marketplace pages
+
+## Back [1.4.1] 2023-08-11
+
+### Fixed
+
+- Fix error : middleware import on users routes
+
+### Changed
+
+_Restructured backend to simplify the logic_
+
+- routes: modified users, removed auth
+- controllers: removed auth and update user, modified search to match new model
+- models: removed all data related model , update model for user and story data
+- middlewares: auth, validations
+- server.js
+
+### Added
+
+- models : storyData
+
+## UI [1.11.0] 2023-08-11
+
+### Changed
+
+- pages: login , register
+- components: userLogin
+- Removed unsused marketplace pages
+
+### Added
+
+- components : userRgister
+
+## Back [1.3.0] 2023-08-10
+
+### Changed
+
+- Routes : auth ,users
+- Models: user
+- Middlewares: auth , validation
+- Controllers: auth , user
+- Server.js : updated routes
+- README file with all required explaination (will need to add later users logins)
+
 ## UI [1.10.0] 2023-08-07
 
 ### Added
+
 - Page: Contact us page
 - CSS: Basic styling added
 
 ## UI [1.9.0] 2023-08-06
 
 ### Added
+
 - Page: Support Page and somee content
 - CSS: Add css page for the support page and added some styling
-
 
 ## UI [1.8.0] 2023-08-01